+++ jacoco-filtering-extension/src/main/kotlin/io/github/surpsg/deltacoverage/jvm/JvmClassUtil.kt
@@ -1,0 +1,21 @@
+package io.github.surpsg.deltacoverage.jvm
+
+import org.objectweb.asm.ClassReader
+import org.objectweb.asm.ClassVisitor
+import org.objectweb.asm.Opcodes
+
+internal fun readClassFileName(classBytes: ByteArray): String? {
+    val customClassVisitor = CustomClassVisitor()
+    ClassReader(classBytes).accept(customClassVisitor, ClassReader.SKIP_CODE)
+    return customClassVisitor.sourceName
+}
+
+private class CustomClassVisitor : ClassVisitor(Opcodes.ASM7) {
+
+    var sourceName: String? = null
+
+    override fun visitSource(source: String?, debug: String?) {
+        super.visitSource(source, debug)
+        sourceName = source
+    }
+}